requirements.txt
# Core GCP APIs
google-cloud-compute>=1.14.0
google-cloud-billing>=1.10.0
google-cloud-storage>=2.10.0

# Templating
jinja2>=3.1.2

# Data Validation & Settings
pydantic>=2.4.2
pydantic-settings>=2.0.0

# CLI Framework
typer[all]>=0.9.0

# Resilience & Retries
tenacity>=8.2.3

# Logging
colorlog>=6.7.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
